Move adc definiteions to init file out of sensor
